add parameter to InputFilesCreator, for creating bidirectional relationships add parameter to ShortestPathBench for selecting Direction of path expander

+++ src/main/java/org/neo4j/bench/shortestpath/InputFilesCreator.java
@@ -11,55 +11,104 @@ import java.util.Set;
 
 public class InputFilesCreator
 {
+    private static String DIRECTED_FALSE = "unidirectional";
+    private static String DIRECTED_TRUE = "bidirectional";
+
     public static void main( String[] args ) throws IOException
     {
+        String errMsg = String.format( "Expected 1 parameter, found %s. Parameters should be <%s|%s>\n", args.length,
+                DIRECTED_FALSE, DIRECTED_TRUE );
+
+        if ( args.length != 1 )
+        {
+            System.out.println( errMsg );
+            return;
+        }
+
+        if ( args[0].equals( DIRECTED_FALSE ) == false && args[0].equals( DIRECTED_TRUE ) == false )
+        {
+            System.out.println( String.format( "Unexpected value for parameter 0: %s\n%s", args[0], errMsg ) );
+            return;
+        }
+
         Random random = new Random( 42 );
         DecimalFormat doubleFormat = new DecimalFormat( "#.##" );
         int pathCount = 10000;
+        boolean directed = ( args[0].equals( DIRECTED_TRUE ) );
 
         File relationshipsCsvFile = new File( Config.RELATIONSHIP_ID_FILE );
         relationshipsCsvFile.delete();
         relationshipsCsvFile.createNewFile();
-        CsvFileWriter relationshipsCsvWriter = new CsvFileWriter( relationshipsCsvFile );
-        relationshipsCsvWriter.writeLine( "from", "to", "type", "weight@double" );
+        CsvFileWriter relationshipsWriter = new CsvFileWriter( relationshipsCsvFile );
+        CsvFileReader relationshipsReader = new CsvFileReader( new File( Config.RAW_RELATIONSHIP_FILE ) );
 
         File nodesCsvFile = new File( Config.NODE_ID_FILE );
         nodesCsvFile.delete();
         nodesCsvFile.createNewFile();
-        CsvFileWriter nodesCsvWriter = new CsvFileWriter( nodesCsvFile );
-        nodesCsvWriter.writeLine( "id", "weight@double" );
+        CsvFileWriter nodesWriter = new CsvFileWriter( nodesCsvFile );
 
         File pathStartAndEndNodesCsvFile = new File( Config.PATH_START_END_ID_FILE );
         pathStartAndEndNodesCsvFile.delete();
         pathStartAndEndNodesCsvFile.createNewFile();
-        CsvFileWriter pathStartAndEndNodesCsvWriter = new CsvFileWriter( pathStartAndEndNodesCsvFile );
-        pathStartAndEndNodesCsvWriter.writeLine( "start", "end" );
+        CsvFileWriter pathStartAndEndNodesWriter = new CsvFileWriter( pathStartAndEndNodesCsvFile );
 
-        // Make Weighted Relationship File
+        Set<Long> nodeIds = createRelationshipFileAndReturnUniqueNodeIds( relationshipsReader, relationshipsWriter,
+                random, doubleFormat, directed );
+
+        createNodeFile( nodeIds, nodesWriter, random, doubleFormat );
+
+        // Make Path Start And End Nodes File
+        createPathStartAndEndNodeFile( pathStartAndEndNodesWriter, nodeIds, pathCount, random );
+
+        relationshipsWriter.close();
+        nodesWriter.close();
+        pathStartAndEndNodesWriter.close();
+    }
+
+    // Make Weighted Relationship File
+    private static Set<Long> createRelationshipFileAndReturnUniqueNodeIds( CsvFileReader relationshipsReader,
+            CsvFileWriter relationshipsWriter, Random random, DecimalFormat doubleFormat, boolean bidirectional )
+            throws IOException
+    {
+        relationshipsWriter.writeLine( "from", "to", "type", "weight@double" );
         Set<Long> nodeIds = new HashSet<Long>();
-        CsvFileReader reader = new CsvFileReader( new File( Config.RAW_RELATIONSHIP_FILE ) );
-        while ( reader.hasNext() )
+        while ( relationshipsReader.hasNext() )
         {
-            String[] relationshipNodes = reader.next();
+            String[] relationshipNodes = relationshipsReader.next();
             for ( String node : relationshipNodes )
             {
                 nodeIds.add( incByOne( node ) );
             }
             double weight = random.nextDouble();
-            relationshipsCsvWriter.writeLine( incByOne( relationshipNodes[0] ).toString(),
-                    incByOne( relationshipNodes[1] ).toString(), "LINK", doubleFormat.format( weight ) );
+            String correctedStartNodeId = incByOne( relationshipNodes[0] ).toString();
+            String correctedEndNodeId = incByOne( relationshipNodes[1] ).toString();
+            String formattedWeight = doubleFormat.format( weight );
+            relationshipsWriter.writeLine( correctedStartNodeId, correctedEndNodeId, "LINK", formattedWeight );
+            if ( bidirectional )
+            {
+                relationshipsWriter.writeLine( correctedEndNodeId, correctedStartNodeId, "LINK", formattedWeight );
+            }
         }
-        relationshipsCsvWriter.close();
+        return nodeIds;
+    }
 
-        // Make Weighted Node File
+    // Make Weighted Node File
+    private static void createNodeFile( Set<Long> nodeIds, CsvFileWriter nodesWriter, Random random,
+            DecimalFormat doubleFormat ) throws IOException
+    {
+        nodesWriter.writeLine( "id", "weight@double" );
         for ( Long nodeId : nodeIds )
         {
             double weight = random.nextDouble();
-            nodesCsvWriter.writeLine( nodeId.toString(), doubleFormat.format( weight ) );
+            nodesWriter.writeLine( nodeId.toString(), doubleFormat.format( weight ) );
         }
-        nodesCsvWriter.close();
+    }
 
-        // Make Path Start And End Nodes File
+    // Make Path Start And End Nodes File
+    private static void createPathStartAndEndNodeFile( CsvFileWriter pathStartAndEndNodesWriter, Set<Long> nodeIds,
+            int pathCount, Random random ) throws IOException
+    {
+        pathStartAndEndNodesWriter.writeLine( "start", "end" );
         List<Long> allNodeIds = Arrays.asList( nodeIds.toArray( new Long[nodeIds.size()] ) );
         int allNodeIdsCount = allNodeIds.size();
         for ( int i = 0; i < pathCount; i++ )
@@ -71,9 +120,8 @@ public class InputFilesCreator
                 long val = allNodeIds.get( (int) ( random.nextDouble() * allNodeIdsCount ) );
                 endNodeId = ( val == startNodeId ) ? -1 : val;
             }
-            pathStartAndEndNodesCsvWriter.writeLine( startNodeId.toString(), endNodeId.toString() );
+            pathStartAndEndNodesWriter.writeLine( startNodeId.toString(), endNodeId.toString() );
         }
-        pathStartAndEndNodesCsvWriter.close();
     }
 
     // To avoid Reference Node in Neo4j, make sure no Nodes have ID==0
